Kalman filter ported to LPC1768
This is the initial port of the Kalman filter from the Processing
orientation.pde sample running on the PC to C++ code running on the
LPC1768 microcontroller. The LPC1768 is currently idle for more than
89% of the time when running this code. This is much better than I
expected for all of that floating point math and gives me lots of
headroom to execute my robot navigation code.

NOTE: This ported code appears to work as expected when switching the
      orientation.pde sample into Embedded mode (by pressing the E
      key). The orientation sample does issue println()s though
      indicating that there are some differences in the PC and embedded
      calculated orientation quaternions until the filter converges. I
      am going to start investigating these diffs as soon as I push up
      this commit. It could be that the PC version gets more precision
      by using special multiply-accumulate or dot-product instructions.

* Added the ability to fetch a float from the config file. Previously
  it could only fetch 3-element vectors of floats or integers.
* Use the new ConfigFile::getFloat() method to fetch co-variance values
  from config.ini to be used by the Kalman filter. The PC version has
  these hardcoded.
* The ported version of the code is no longer hard-coded to run at
  100Hz. The Sparkfun9DoFSensorStick constructor now takes this refresh
  rate as a parameter that defaults to 100Hz.
* Updated an assert() in FlashFileSystem.cpp to not issue a signed/
  unsigned comparison warning during Debug builds.
* Add a Matrix4x4 class which was written specifically for use with
  this Kalman filter. It only supports the operations required by the
  code for the Kalman filter. It also contains optimized methods such
  as multiplyTransposed() and addDiagonal(). These reduce the work
  load by getting rid of the need to do things like transposing a
  matrix for the sole purpose of a subsequent multiplication or adding
  zeroes to most elements in a matrix when it is known that one of the
  matrices only contains non-zero values on the diagonal.
* The Marix4x4 class takes advantage of ARM's CMSIS-DSP library for
  calculating the inverse of a 4x4 matrix.
* There are other operations like matrix multiplication and addition
  that are supported by the CMSIS-DSP library but they weren't
  utilized. Instead inline compatible versions were provided in the
  Matrix4x4 class header. These methods are specific to 4x4 matrices so
  they don't need to be generalized for matrices of any size and the
  loops could be explicitly unrolled to improve performance. Making
  these switches away from CMSIS-DSP general routines resulted in a 10%
  performance improvement for the running the Kalman filter.
* Added a few more methods to the Quaternion class that were required
  for calculating the Kalman filter.
* Updating the orientation.pde Processing sample to reset the Kalman
  filter state when it has detected that the embedded version has done
  the same. It was also updated to expect that the embedded code is now
  running the full Kalman filter and use that knowledge when comparing
  the orientation results.

+++ firmware/FlashFileSystem.cpp
@@ -356,7 +356,7 @@ void FlashFileSystemDirHandle::seekdir(off_t Location)
     
     assert ( NULL != pLocation &&
              pLocation > m_pFirstFileEntry &&
-             (pLocation - m_pFirstFileEntry) < m_FileEntriesLeft );
+             (unsigned int)(pLocation - m_pFirstFileEntry) < m_FileEntriesLeft );
     
     m_pCurrentFileEntry = pLocation;
 }
